+++ R/cohort_builders.R
@@ -376,6 +376,7 @@ get_custom_derived_sql_params <- function(sqlite_conn, cohort_id) {
   return(custom_params)
 }
 
+
 generate_single_cohort <- function(
     cohort, 
     cohort_id,